- realized the server is using int's for the post and user id's, so changing them to int's on the client as well

+++ Rove/src/com/fourkins/rove/activity/NewPostActivity.java
@@ -188,7 +188,7 @@ public class NewPostActivity extends Activity {
         }
     }
 
-    private long getUserId() {
+    private int getUserId() {
         Rove rove = (Rove) getApplication();
         return rove.getUserId();
     }
